Securing Static Files
> **tl;dr:** This patch adds a plug-in based authorization mechanism for
> delivering static files (e.g. videos) which, by default, requires no
> configuration and which adheres to Opencast's default access control
> mechanisms (users, roles, groups, and access control lists).

Opencast comes with a complex access management system with fine-grained
control over users, roles, groups of roles and access rights. But all of
this is blatantly ignored when it comes to serving static file content
like the video files. These are served by Opencast's static module which
is nothing but a file server.

Right now, these file's protection is the form of the URLs which
generally contain at least two UUIDs and is thus hard to guess
(comparable e.g. to the “Share via link” options in Google docs, …). The
problem of this is obviously that if an authorized user publishes the
direct URL to a file, all protection is lost.

To address this problem, Opencast has an implementation for a
token-based authorization system which can be configured to only allow
access with valid tokens which in turn are generated by a trusted entity
and which are valid only for a certain amount of time. The advantage of
this is that it can be used to have external systems authorize
downloads. The disadvantage is that it's hard to configure correctly,
that yet another authorization mechanism ties in the already existing
one and that there are problems with either granting or preventing
access in some edge cases. Can't we do better?

An intuitive solution for this problem would be that the file server
simply adheres to the access control lists set for a given media package
and only delivers files to users which are authorized to access this
content as part of their user rights and roles. So, let's do that.

The Problem
-----------

Opencast's static files server delivers files for all publication
channels. By default these are:

- /static/**api**/<media-package-id>/…
- /static/**engage-player**/<media-package-id>/…
- /static/**internal**/<media-package-id>/…
- /static/**oaipmh-default**/<media-package-id>/…

In other words, several components of Opencast are involved in this and
the module “static” does not know about them. It is not even guaranteed
that the form of URLs is always the same. The module could be used to
serve any path as long as the URL starts with `/static/`.

Solution
--------

To allow for flexibility, this patch allows arbitrary services to
optionally register themselves with “static”, providing an authorization
interface which contains:

- A list of URL patterns, the authorization plug-in feels responsible
  for (e.g. the search service could offer responsibility for
  `/static/engage-player/…`)

- A method checking authorization based on a given URL (e.g. the search
  service would check access based on the media package identifier
  contained in the URL)

Delivering files for URLs which are not allowed by anyone would be
forbidden.

This allows for intuitive use of the authorization mechanisms already in
place for parts of the system without the need to do any complicated,
additional configuration. For example, we publish access control lists
to the engage components anyway and they are used for checking access to
content in the search service based on the user session. Expecting the
same mechanism for content as well would be just natural.

Compatibility
-------------

Does this break the current usage? In most cases, it shouldn't since a
lot of people already rely on the access controls set by Opencast and
other users or systems shouldn't access any files right now.

Nevertheless, there might be the usual special case for which the system
should be easy to turn off, restoring the current behavior of Opencast.
The same can also be achieved by serving the static content directly via
an HTTP server. By default, however, the protection should be turned on
in the long run since this is like the more intuitive understanding when
it comes to protecting content.

An additional way to deal with custom downloads, e.g. for custom
publication channels, would be to provide a simple authorization plug-in
with configurable patterns which can be used to allow additional content
to be downloaded without any specific checks.

Roadmap
-------

This patch targets Opencast 9.2. To ensure a smooth upgrade path and
compatibility to previous 9.x releases, all checks are still off by
default. This means there is no difference running Opencast with or
without this patch in its default configuration.

But to meet users expectation when it comes to securing those files, the
mechanism should be turned on by default in Opencast 10.

For now, turning on the protection can be done in
`etc/org.opencastproject.fsresources.StaticResourceServlet.cfg`